+++ src/petals/server/server.py
@@ -18,7 +18,7 @@ from hivemind.proto.runtime_pb2 import CompressionType
 from hivemind.utils.logging import get_logger, use_hivemind_log_handler
 
 from petals.bloom.from_pretrained import DTYPE_MAP, load_pretrained_block
-from petals.bloom.model import BloomConfig
+from petals.bloom.modeling_utils import BloomConfig
 from petals.constants import PUBLIC_INITIAL_PEERS
 from petals.data_structures import CHAIN_DELIMITER, UID_DELIMITER, ServerState
 from petals.dht_utils import declare_active_modules, get_remote_module_infos
